+++ app/src/components/ProofTree/components/BoxEl/services/prettifyGoalUsername.ts
